requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
factory-boy==3.2.1
Faker==13.11.1
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Mako==1.2.0
MarkupSafe==2.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.36
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.2
yarl==1.7.2
